ETL/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.6
kafka-python==2.0.2
pycryptodome==3.19.0
pymongo==4.6.1
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0
websocket-client==1.7.0
